Revert a choice made in f19787667f28d0d505d42ad63fae74cb9f0e3dd4 that set the 1st dark frame equal to the 0th in the event that only one dark frame was taken, since this misrepresented the data. Update the quicklook plotting code to show the raw data for light frames 1 through n (that go with the second dark frame), rather than the raw data with the 0th dark frame subtracted (incorrect usage).
Explanation: The choice to copy the 0th dark frame into the 1st was originally made because the way I wrote the quicklook code would gray out light frames from 1--n (0-indexed of course) since the dark was bad. This update solves the problem of how to display the light frames based on what their dark frame is doing in the place it occurs--in the quicklook creation code--rather than deeper down in get_dark_frames() which can be called by other functions and could lead to incorrect analyses.